znr: add NULL check for fs ops
The fs->ops can be NULL if we had an early exit. So make sure to check
it first.

+++ src/znr_fs.c
@@ -72,7 +72,7 @@ static int znr_fs_init_fs(struct znr_fs_file *f)
 
 static int znr_fs_destroy_fs(struct znr_fs_file *f)
 {
-	if (f->fs->ops->destroy_fs)
+	if (f->fs->ops && f->fs->ops->destroy_fs)
 		return f->fs->ops->destroy_fs();
 
 	return 0;
